requirements.txt
Flask
PyYAML
jinja2
click
flask-misaka
